changing file structure to lerna / workspaces
