+++ jest.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   "roots": [
+    "./packages",
     "./tests"
   ],
   "transform": {
