+++ package.json
@@ -39,14 +39,11 @@
   ],
   "license": "BSD-3-Clause",
   "dependencies": {},
-  "devDependencies": {
-    "@types/jest": "^24.0.15",
-    "jest": "^24.8.0",
-    "ts-jest": "^24.0.2",
-    "tslint": "^5.18.0",
-    "typescript": "^3.5.3"
-  },
+  "devDependencies": {},
   "workspaces": [
-    "common"
+    "packages/*"
+  ],
+  "files": [
+    "dist"
   ]
 }
