+++ packages/core/abstracts.ts
