+++ packages/core/errors.ts
